backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pypdf
google-generativeai
python-dotenv
pydantic
httpx
